REPO-3684: Renditions: Renditions triggered via Content update policy raw events    - refactored event listener and removed it from RenditionService2Impl
