+++ src/main/java/org/alfresco/messaging/camel/routes/OnContentUpdateRenditionRoute.java
@@ -28,9 +28,21 @@ package org.alfresco.messaging.camel.routes;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import org.alfresco.model.RenditionModel;
+import org.alfresco.repo.content.ContentServicePolicies;
+import org.alfresco.repo.policy.Behaviour;
+import org.alfresco.repo.policy.EventBehaviour;
+import org.alfresco.repo.policy.PolicyComponent;
+import org.alfresco.repo.rawevents.TransactionAwareEventProducer;
+import org.alfresco.repo.rawevents.types.EventType;
+import org.alfresco.repo.rawevents.types.OnContentUpdatePolicyEvent;
+import org.alfresco.repo.security.authentication.AuthenticationUtil;
+import org.alfresco.service.cmr.repository.NodeRef;
+import org.alfresco.util.GUID;
 import org.apache.camel.spring.SpringRouteBuilder;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -50,6 +62,12 @@ public class OnContentUpdateRenditionRoute extends SpringRouteBuilder
     // Not restricted for now, should be restricted after performance tests.
     private ExecutorService executorService = Executors.newCachedThreadPool();
 
+    @Autowired
+    private TransactionAwareEventProducer transactionAwareEventProducer;
+
+    @Autowired
+    private PolicyComponent policyComponent;
+
     @Override
     public void configure() throws Exception
     {
@@ -59,6 +77,29 @@ public class OnContentUpdateRenditionRoute extends SpringRouteBuilder
             logger.debug("SourceQueue is " + sourceQueue);
         }
 
+        EventBehaviour eventBehaviour = new EventBehaviour(transactionAwareEventProducer, sourceQueue, this, "createOnContentUpdateEvent",
+                Behaviour.NotificationFrequency.EVERY_EVENT);
+        policyComponent.bindClassBehaviour(ContentServicePolicies.OnContentUpdatePolicy.QNAME, RenditionModel.ASPECT_RENDITIONED, eventBehaviour);
+
         from(sourceQueue).threads().executorService(executorService).process("renditionEventProcessor").end();
+    }
+
+    @SuppressWarnings("unused")
+    public OnContentUpdatePolicyEvent createOnContentUpdateEvent(NodeRef sourceNodeRef, boolean newContent)
+    {
+        OnContentUpdatePolicyEvent event = new OnContentUpdatePolicyEvent();
+
+        // Raw event specific
+        event.setId(GUID.generate());
+        event.setType(EventType.CONTENT_UPDATED.toString());
+        event.setAuthenticatedUser(AuthenticationUtil.getFullyAuthenticatedUser());
+        event.setExecutingUser(AuthenticationUtil.getRunAsUser());
+        event.setTimestamp(System.currentTimeMillis());
+        event.setSchema(1);
+
+        // On content update policy event specific
+        event.setNodeRef(sourceNodeRef.toString());
+        event.setNewContent(newContent);
+        return event;
     }
 }
