+++ src/main/java/org/alfresco/repo/rendition2/RenditionService2Impl.java
@@ -29,13 +29,8 @@ import org.alfresco.model.ContentModel;
 import org.alfresco.model.RenditionModel;
 import org.alfresco.repo.content.ContentServicePolicies;
 import org.alfresco.repo.content.MimetypeMap;
-import org.alfresco.repo.policy.Behaviour;
 import org.alfresco.repo.policy.BehaviourFilter;
-import org.alfresco.repo.policy.EventBehaviour;
 import org.alfresco.repo.policy.PolicyComponent;
-import org.alfresco.repo.rawevents.AbstractEventProducer;
-import org.alfresco.repo.rawevents.types.EventType;
-import org.alfresco.repo.rawevents.types.OnContentUpdatePolicyEvent;
 import org.alfresco.repo.rendition.RenditionPreventionRegistry;
 import org.alfresco.repo.security.authentication.AuthenticationUtil;
 import org.alfresco.repo.transaction.RetryingTransactionHelper;
@@ -53,7 +48,6 @@ import org.alfresco.service.namespace.NamespaceService;
 import org.alfresco.service.namespace.QName;
 import org.alfresco.service.namespace.RegexQNamePattern;
 import org.alfresco.service.transaction.TransactionService;
-import org.alfresco.util.GUID;
 import org.alfresco.util.PropertyCheck;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -101,8 +95,6 @@ public class RenditionService2Impl implements RenditionService2, InitializingBea
     private BehaviourFilter behaviourFilter;
     private RuleService ruleService;
     private PostTxnCallbackScheduler renditionRequestSheduler;
-    private AbstractEventProducer eventProducer;
-    private String rawEventsEndpoint;
     private boolean enabled;
     private boolean thumbnailsEnabled;
 
@@ -162,14 +154,6 @@ public class RenditionService2Impl implements RenditionService2, InitializingBea
         this.ruleService = ruleService;
     }
 
-    public void setEventProducer(AbstractEventProducer eventProducer) {
-        this.eventProducer = eventProducer;
-    }
-
-    public void setRawEventsEndpoint(String rawEventsEndpoint) {
-        this.rawEventsEndpoint = rawEventsEndpoint;
-    }
-
     public void setEnabled(boolean enabled)
     {
         this.enabled = enabled;
@@ -197,9 +181,6 @@ public class RenditionService2Impl implements RenditionService2, InitializingBea
         PropertyCheck.mandatory(this, "policyComponent", policyComponent);
         PropertyCheck.mandatory(this, "behaviourFilter", behaviourFilter);
         PropertyCheck.mandatory(this, "ruleService", ruleService);
-
-        EventBehaviour eventBehaviour = new EventBehaviour(eventProducer, rawEventsEndpoint, this, "createOnContentUpdateEvent", Behaviour.NotificationFrequency.EVERY_EVENT);
-        policyComponent.bindClassBehaviour(ContentServicePolicies.OnContentUpdatePolicy.QNAME, RenditionModel.ASPECT_RENDITIONED, eventBehaviour);
     }
 
     public void render(NodeRef sourceNodeRef, String renditionName)
@@ -682,24 +663,6 @@ public class RenditionService2Impl implements RenditionService2, InitializingBea
                 }
             }
         }
-    }
-
-    public OnContentUpdatePolicyEvent createOnContentUpdateEvent(NodeRef sourceNodeRef, boolean newContent)
-    {
-        OnContentUpdatePolicyEvent event = new OnContentUpdatePolicyEvent();
-
-        // Raw event specific
-        event.setId(GUID.generate());
-        event.setType(EventType.CONTENT_UPDATED.toString());
-        event.setAuthenticatedUser(AuthenticationUtil.getFullyAuthenticatedUser());
-        event.setExecutingUser(AuthenticationUtil.getRunAsUser());
-        event.setTimestamp(System.currentTimeMillis());
-        event.setSchema(1);
-
-        // On content update policy event specific
-        event.setNodeRef(sourceNodeRef.toString());
-        event.setNewContent(newContent);
-        return event;
     }
 
 }
