+++ src/main/resources/alfresco/rendition-services2-context.xml
@@ -59,8 +59,6 @@
         <property name="policyComponent" ref="policyComponent" />
         <property name="behaviourFilter" ref="policyBehaviourFilter" />
         <property name="ruleService" ref="ruleService" />
-        <property name="eventProducer" ref="transactionAwareEventProducer" />
-        <property name="rawEventsEndpoint" value="${acs.repo.rendition.events.endpoint}" />
         <property name="enabled" value="${renditionService2.enabled}" />
         <property name="thumbnailsEnabled" value="${system.thumbnail.generate}" />
         <property name="renditionRequestSheduler" ref="renditionRequestSheduler" />
